Refactor: Move components to different folders
